Create independent subclass tests
